apps/api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
httpx==0.27.0
supabase==2.7.0
asyncpg==0.29.0
pydantic==2.9.0
pydantic-settings==2.5.0
google-generativeai==0.8.0
redis==5.0.0
